+++ mobile/android/android-components/components/feature/syncedtabs/src/main/java/mozilla/components/feature/syncedtabs/presenter/DefaultPresenter.kt
@@ -81,7 +81,8 @@ internal class DefaultPresenter(
     }
 
     override fun stop() {
-        // no-op
+        accountManager.unregisterForSyncEvents(eventObserver)
+        accountManager.unregister(accountObserver)
     }
 
     companion object {